Soln03_Zephyr: Blink LED with delay time received over UART
This patch is inspired from challenge at www.digiKey.in
(https://www.digikey.in/en/maker/projects/introduction-to
-rtos-solution-to-part-3-task-scheduling/8fbb9e0b0eed4279a2dd698f02ce125f)

Challenge:
Using FreeRTOS, create two separate tasks. One listens for
an integer over UART (from the Serial Monitor) and sets a
variable when it sees an integer. The other task blinks the
onboard LED (or other connected LED) at a rate specified by
that integer. In effect, you want to create a multi-threaded
system that allows for the user interface to run concurrently
with the control task (the blinking LED).

Implement on Zephyr OS for ESP32, nrf52840.
The 'uart0' serial port  is aliased as 'usercom0' in the device
tree overlay.