Soln06_Zephyr: Demo use of mutex, semaphore (kernel mode)
This patch is inspired from challenge at  www.digiKey.in
(https://www.digikey.in/en/maker/projects/introduction-to-rtos
-solution-to-part-6-freertos-mutex-example/c6e3581aa2204f1380e83a9b4c3807a6)

The Challenge aims to demo the passing of parameters to the task,
its unintended effects and solution to overcome the issue.

In this patch, use mutex to protect led instance for toggling
and semaphore to signal parameters read for thread 0.
This patch also enables module instance logging for led module.