Soln12FreeRTOS: Multicore Systems Challenge
This patch is inspired from challenge at www.digiKey.in
(https://www.digikey.in/en/maker/projects/introduction-to-
rtos-solution-to-part-12-multicore-systems/369936f5671d4207a2c954c0637e7d50)

Challenge:

Your challenge is to start with your hardware interrupt solution
and make it run with dual-core support on the ESP32. Specifically,
the timer ISR (for sampling the ADC) and the task that computes the
average of 10 samples should run on Core 0. The user interface task
(the one that handles echoing characters to the serial terminal and
responding to the “avg” command) should run on Core 1.

FIXME: Though the assignment of uart thread to CPU core 1 is sucessful
in ESP32, the thread is not started.